+++ packages/paste-core/components/radio-group/src/Radio.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import * as PropTypes from 'prop-types';
-import {useUID} from 'react-uid';
+import {useUID} from '@twilio-paste/uid-library';
 import {Box} from '@twilio-paste/box';
 import {
   BaseRadioCheckboxControl,