+++ packages/paste-core/components/toast/src/Toast.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import {Button} from '@twilio-paste/button';
 import {CloseIcon} from '@twilio-paste/icons/esm/CloseIcon';
 import {ErrorIcon} from '@twilio-paste/icons/esm/ErrorIcon';
-import {InformationIcon} from '@twilio-paste/icons/esm/InformationIcon';
+import {NeutralIcon} from '@twilio-paste/icons/esm/NeutralIcon';
 import {SuccessIcon} from '@twilio-paste/icons/esm/SuccessIcon';
 import {WarningIcon} from '@twilio-paste/icons/esm/WarningIcon';
 import {MediaObject, MediaFigure, MediaBody} from '@twilio-paste/media-object';
@@ -30,31 +30,40 @@ const renderToastIcon = (variant: ToastVariants): React.ReactElement => {
       return <WarningIcon color="colorTextWarning" decorative={false} title="warning: " size="sizeIcon20" />;
     case ToastVariantObject.NEUTRAL:
     default:
-      return <InformationIcon color="colorTextNeutral" decorative={false} title="information: " size="sizeIcon20" />;
+      return <NeutralIcon color="colorTextNeutral" decorative={false} title="information: " size="sizeIcon20" />;
   }
 };
 
-const Toast: React.FC<ToastProps> = ({children, onDismiss, variant = 'neutral', ...props}) => {
-  const ToastComponent = ToastComponentVariants[variant];
+const Toast = React.forwardRef<HTMLDivElement, ToastProps>(
+  ({children, onDismiss, variant = 'neutral', setFocus, ...props}, ref) => {
+    const ToastComponent = ToastComponentVariants[variant];
+    const buttonRef: React.RefObject<HTMLButtonElement> = React.useRef(null);
 
-  return (
-    <ToastComponent role="status" variant={variant} {...props}>
-      <MediaObject as="div">
-        <MediaFigure as="div" spacing="space60">
-          {renderToastIcon(variant)}
-        </MediaFigure>
-        <MediaBody as="div">{children}</MediaBody>
-        {onDismiss && typeof onDismiss === 'function' && (
-          <MediaFigure align="end" spacing="space40">
-            <Button onClick={onDismiss} variant="link" size="reset">
-              <CloseIcon color="colorTextIcon" decorative={false} title="dismiss this toast" size="sizeIcon20" />
-            </Button>
+    React.useEffect(() => {
+      if (setFocus && buttonRef.current) {
+        buttonRef.current.focus({preventScroll: true});
+      }
+    }, [setFocus]);
+
+    return (
+      <ToastComponent role="status" variant={variant} ref={ref} {...props}>
+        <MediaObject as="div">
+          <MediaFigure as="div" spacing="space60">
+            {renderToastIcon(variant)}
           </MediaFigure>
-        )}
-      </MediaObject>
-    </ToastComponent>
-  );
-};
+          <MediaBody as="div">{children}</MediaBody>
+          {onDismiss && typeof onDismiss === 'function' && (
+            <MediaFigure align="end" spacing="space40">
+              <Button onClick={onDismiss} variant="link" ref={buttonRef} size="reset">
+                <CloseIcon color="colorTextIcon" decorative={false} title="dismiss this toast" size="sizeIcon20" />
+              </Button>
+            </MediaFigure>
+          )}
+        </MediaObject>
+      </ToastComponent>
+    );
+  }
+);
 
 Toast.displayName = 'Toast';