+++ packages/paste-libraries/animation/rollup.config.js
@@ -27,6 +27,7 @@ export default {
       tsconfig: './tsconfig.build.json',
     }),
     babel({
+      runtimeHelpers: true,
       exclude: 'node_modules/**',
     }),
     process.env.NODE_ENV === 'production' ? terser({output: {comments: false}}) : null,