+++ packages/paste-core/components/anchor/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.11](https://github.com/twilio-labs/paste/compare/@twilio-paste/anchor@1.3.10...@twilio-paste/anchor@1.3.11) (2020-10-21)
+
+**Note:** Version bump only for package @twilio-paste/anchor
+
+
+
+
+
+## [1.3.10](https://github.com/twilio-labs/paste/compare/@twilio-paste/anchor@1.3.9...@twilio-paste/anchor@1.3.10) (2020-10-19)
+
+**Note:** Version bump only for package @twilio-paste/anchor
+
+
+
+
+
+## [1.3.9](https://github.com/twilio-labs/paste/compare/@twilio-paste/anchor@1.3.8...@twilio-paste/anchor@1.3.9) (2020-10-15)
+
+**Note:** Version bump only for package @twilio-paste/anchor
+
+
+
+
+
 ## [1.3.8](https://github.com/twilio-labs/paste/compare/@twilio-paste/anchor@1.3.7...@twilio-paste/anchor@1.3.8) (2020-10-13)
 
 **Note:** Version bump only for package @twilio-paste/anchor