+++ packages/paste-core/components/combobox/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.11](https://github.com/twilio-labs/paste/compare/@twilio-paste/combobox@1.4.10...@twilio-paste/combobox@1.4.11) (2020-10-22)
+
+**Note:** Version bump only for package @twilio-paste/combobox
+
+
+
+
+
+## [1.4.10](https://github.com/twilio-labs/paste/compare/@twilio-paste/combobox@1.4.9...@twilio-paste/combobox@1.4.10) (2020-10-21)
+
+**Note:** Version bump only for package @twilio-paste/combobox
+
+
+
+
+
+## [1.4.9](https://github.com/twilio-labs/paste/compare/@twilio-paste/combobox@1.4.8...@twilio-paste/combobox@1.4.9) (2020-10-19)
+
+**Note:** Version bump only for package @twilio-paste/combobox
+
+
+
+
+
+## [1.4.8](https://github.com/twilio-labs/paste/compare/@twilio-paste/combobox@1.4.7...@twilio-paste/combobox@1.4.8) (2020-10-15)
+
+**Note:** Version bump only for package @twilio-paste/combobox
+
+
+
+
+
 ## [1.4.7](https://github.com/twilio-labs/paste/compare/@twilio-paste/combobox@1.4.6...@twilio-paste/combobox@1.4.7) (2020-10-13)