+++ packages/paste-core/components/inline-control-group/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.8](https://github.com/twilio-labs/paste/compare/@twilio-paste/inline-control-group@1.1.7...@twilio-paste/inline-control-group@1.1.8) (2020-10-21)
+
+**Note:** Version bump only for package @twilio-paste/inline-control-group
+
+
+
+
+
+## [1.1.7](https://github.com/twilio-labs/paste/compare/@twilio-paste/inline-control-group@1.1.6...@twilio-paste/inline-control-group@1.1.7) (2020-10-19)
+
+**Note:** Version bump only for package @twilio-paste/inline-control-group
+
+
+
+
+
+## [1.1.6](https://github.com/twilio-labs/paste/compare/@twilio-paste/inline-control-group@1.1.5...@twilio-paste/inline-control-group@1.1.6) (2020-10-15)
+
+**Note:** Version bump only for package @twilio-paste/inline-control-group
+
+
+
+
+
 ## [1.1.5](https://github.com/twilio-labs/paste/compare/@twilio-paste/inline-control-group@1.1.4...@twilio-paste/inline-control-group@1.1.5) (2020-10-13)
 
 **Note:** Version bump only for package @twilio-paste/inline-control-group