+++ packages/paste-core/components/label/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.6](https://github.com/twilio-labs/paste/compare/@twilio-paste/label@1.1.5...@twilio-paste/label@1.1.6) (2020-10-21)
+
+**Note:** Version bump only for package @twilio-paste/label
+
+
+
+
+
+## [1.1.5](https://github.com/twilio-labs/paste/compare/@twilio-paste/label@1.1.4...@twilio-paste/label@1.1.5) (2020-10-15)
+
+**Note:** Version bump only for package @twilio-paste/label
+
+
+
+
+
 ## [1.1.4](https://github.com/twilio-labs/paste/compare/@twilio-paste/label@1.1.3...@twilio-paste/label@1.1.4) (2020-10-13)
 
 **Note:** Version bump only for package @twilio-paste/label