+++ packages/paste-core/components/typography/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.82](https://github.com/twilio-labs/paste/compare/@twilio-paste/typography@1.0.81...@twilio-paste/typography@1.0.82) (2020-10-15)
+
+**Note:** Version bump only for package @twilio-paste/typography
+
+
+
+
+
 ## [1.0.81](https://github.com/twilio-labs/paste/compare/@twilio-paste/typography@1.0.80...@twilio-paste/typography@1.0.81) (2020-10-13)
 
 **Note:** Version bump only for package @twilio-paste/typography