+++ packages/paste-core/core-bundle/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.6](https://github.com/twilio-labs/paste/compare/@twilio-paste/core@3.0.5...@twilio-paste/core@3.0.6) (2020-10-22)
+
+**Note:** Version bump only for package @twilio-paste/core
+
+
+
+
+
+## [3.0.5](https://github.com/twilio-labs/paste/compare/@twilio-paste/core@3.0.4...@twilio-paste/core@3.0.5) (2020-10-21)
+
+**Note:** Version bump only for package @twilio-paste/core
+
+
+
+
+
+## [3.0.4](https://github.com/twilio-labs/paste/compare/@twilio-paste/core@3.0.3...@twilio-paste/core@3.0.4) (2020-10-19)
+
+**Note:** Version bump only for package @twilio-paste/core
+
+
+
+
+
+## [3.0.3](https://github.com/twilio-labs/paste/compare/@twilio-paste/core@3.0.2...@twilio-paste/core@3.0.3) (2020-10-15)
+
+**Note:** Version bump only for package @twilio-paste/core
+
+
+
+
+
 ## [3.0.2](https://github.com/twilio-labs/paste/compare/@twilio-paste/core@3.0.1...@twilio-paste/core@3.0.2) (2020-10-13)
 
 **Note:** Version bump only for package @twilio-paste/core